% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{default_label}
\alias{default_label}
\title{What is the default label for a plot aesthetic?}
\usage{
default_label(p, aes = NULL)
}
\arguments{
\item{p}{A \link[ggplot2:ggplot]{ggplot} object}

\item{aes}{If \code{aes} is a \link{character} vector, returns only the labels
corresponding to the included aesthetics.
Defaults to \code{\link{NULL}}, which returns all labels.}
}
\value{
A named \link{list} in which each element is a \link{character} string
or \code{\link{NULL}}
}
\description{
What is the default label for a plot aesthetic?
}
\examples{
require(ggplot2)

p <- ggplot(data = mpg, mapping = aes(x = displ, y = hwy)) +
  geom_point(mapping = aes(color = class, shape = drv)) +
  geom_smooth() +
  labs(title = "My plot", x = "Weight", y = "MPG", color = NULL)

# Returns the label the ggplot would create by default for an aesthetic
default_label(p, "x")
default_label(p, "x", "y")
default_label(p)

# If an aesthetic does not exist, returns NULL
default_label(p, "size")

# If an aesthetic has no default, returns NULL
default_label(p, "title")

# The colo(u)r aesthetic can be matched with or without a u
default_label(p, "color")
default_label(p, "colour")
}
\seealso{
Other functions for checking labels: 
\code{\link{get_labels}()},
\code{\link{uses_labels}()}
}
\concept{functions for checking labels}
