% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_params.R
\name{default_params}
\alias{default_params}
\title{What are the default parameters for a plot layer?}
\usage{
default_params(p, geom, ..., i = NULL)
}
\arguments{
\item{p}{A ggplot object}

\item{geom}{A character string found in the suffix of a ggplot2 geom function,
e.g. \code{"point"}.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}>
\link[=character]{Character} strings.
\code{default_params()} returns the default parameter value with a name matching
each string passed to \code{...}.
If no arguments are passed to \code{...} (the default), the default values for
all parameters are returned.}

\item{i}{A numerical index, e.g. \code{1}.}
}
\value{
A named \link{list} of the same length as the number of inputs to \code{...},
or, if \code{...} is empty, a named list of default values for all parameters
of \code{geom}.
}
\description{
What are the default parameters for a plot layer?
}
\examples{
require(ggplot2)

p <- ggplot(data = mpg, mapping = aes(x = displ, y = hwy)) +
  geom_smooth(aes(color = class))

# Returns the parameters the ggplot would use by default for a layer
default_params(p, "smooth", "linetype")
default_params(p, "smooth", "se", "level")
default_params(p, "smooth")

# If a parameter does not exist, returns NULL
default_params(p, "smooth", "shape")

# The colo(u)r aesthetic can be matched with or without a u
default_params(p, "smooth", "color")
default_params(p, "smooth", "colour")
}
\seealso{
Other functions for checking geom parameters: 
\code{\link{uses_geom_params}()}
}
\concept{functions for checking geom parameters}
