% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geoms_match}
\alias{geoms_match}
\title{Does a plot use an exact set of geoms?}
\usage{
geoms_match(p, geoms)
}
\arguments{
\item{p}{A ggplot object}

\item{geoms}{A vector of character strings. Each element should correspond to
the suffix of a ggplot2 \code{geom_} function, e.g. \code{c("point",
  "line", "smooth")}.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
\code{geoms_match} tests whether the geoms used by a plot exactly match the
set of geoms described by \code{geoms} in both order and type.
}
\examples{
p <- ggplot(data = mpg, mapping = aes(x = displ, y = hwy)) +
  geom_point(mapping = aes(color = class)) +
  geom_smooth()
geoms_match(p, geoms = c("point", "smooth"))
}
\seealso{
Other functions for checking geoms: 
\code{\link{get_geoms}()},
\code{\link{ith_geom_is}()},
\code{\link{ith_geom}()},
\code{\link{uses_geoms}()}
}
\concept{functions for checking geoms}
