% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{get_coordinate_system}
\alias{get_coordinate_system}
\title{Which coordinate system does a plot use?}
\usage{
get_coordinate_system(p)
}
\arguments{
\item{p}{A ggplot2 object}
}
\value{
A character string that corresponds to the suffix of a ggplot2
\code{coord_} function, e.g. \code{"cartesian"}.
}
\description{
Which coordinate system does a plot use?
}
\examples{
p <- ggplot(data = mpg, mapping = aes(x = displ, y = hwy)) +
  geom_point(mapping = aes(color = class)) +
  geom_smooth() +
  coord_polar()
get_coordinate_system(p)
}
\seealso{
Other functions for checking coordinate systems: 
\code{\link{uses_coordinate_system}()}
}
\concept{functions for checking coordinate systems}
