% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{get_geoms}
\alias{get_geoms}
\title{List the geoms used by a plot}
\usage{
get_geoms(p)
}
\arguments{
\item{p}{A ggplot object}
}
\value{
A vector of character strings. Each element corresponds to the suffix
of a ggplot2 \code{geom_} function, e.g. \code{c("point", "line", "smooth")}.
}
\description{
\code{get_geoms} returns a vector of geom names, written as character
strings, that describes which geoms in which order are used by a plot.
}
\examples{
p <- ggplot(data = mpg, mapping = aes(x = displ, y = hwy)) +
  geom_point(mapping = aes(color = class)) +
  geom_smooth()
get_geoms(p)
}
\seealso{
Other functions for checking geoms: 
\code{\link{geoms_match}()},
\code{\link{ith_geom_is}()},
\code{\link{ith_geom}()},
\code{\link{uses_geoms}()}
}
\concept{functions for checking geoms}
