% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{get_geoms_stats}
\alias{get_geoms_stats}
\title{List the geom and stat combination used by all layers of a plot.}
\usage{
get_geoms_stats(p)
}
\arguments{
\item{p}{A ggplot object}
}
\value{
A list of lists with a GEOM and STAT character.
e.g. list(list(GEOM = "point", STAT = "identity"))
}
\description{
List the geom and stat combination used by all layers of a plot.
}
\examples{
require(ggplot2)
p <- ggplot(data = mpg, mapping = aes(x = displ, y = hwy)) +
  geom_point(mapping = aes(color = class)) +
  geom_smooth()
get_geoms_stats(p)
}
\seealso{
Other functions for checking geoms: 
\code{\link{get_geoms}()},
\code{\link{ith_geom_is}()},
\code{\link{ith_geom_stat}()},
\code{\link{ith_geom}()},
\code{\link{uses_geoms}()}
}
\concept{functions for checking geoms}
