% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{get_labels}
\alias{get_labels}
\title{List the labels used by a plot}
\usage{
get_labels(p, aes = NULL)
}
\arguments{
\item{p}{A \link[ggplot2:ggplot]{ggplot} object}

\item{aes}{If \code{aes} is a \link{character} vector, returns only the labels
corresponding to the included aesthetics.
Defaults to \code{\link{NULL}}, which returns all labels.}
}
\value{
A named list of character strings.
}
\description{
\code{get_labels()} returns a named \link{list} of \link[ggplot2:labs]{labels},
written as \link{character} strings, indicating which labels are used by a plot.
}
\details{
Note that \code{get_labels()} will return \code{\link{NULL}} if a label is explicitly set to
\code{\link{NULL}} \emph{\strong{or}} if a requested aesthetic is not present in the plot.
}
\examples{
require(ggplot2)

p <- ggplot(data = mpg, mapping = aes(x = displ, y = hwy)) +
  geom_point(mapping = aes(color = class)) +
  geom_smooth() +
  labs(x = "Weight", y = "MPG", color = NULL)

get_labels(p)
get_labels(p, c("x", "y"))

# The colo(u)r aesthetic can be matched with or without a u
get_labels(p, "color")
get_labels(p, "colour")
}
\seealso{
Other functions for checking labels: 
\code{\link{default_label}()},
\code{\link{uses_labels}()}
}
\concept{functions for checking labels}
