% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{get_mappings}
\alias{get_mappings}
\title{Get aesthetic mappings from a layer or plot}
\usage{
get_mappings(p, local_only = FALSE)
}
\arguments{
\item{p}{A ggplot object or a layer extracted from a ggplot object with
\code{\link{get_geom_layer}} or \code{\link{get_stat_layer}}.}

\item{local_only}{\code{TRUE} or \code{FALSE}. Should \code{get_mappings}
return only the mappings defined locally in a layer. This has no effect
when \code{p} is a ggplot object.}
}
\value{
A list with class uneval, as returned by \code{\link[ggplot2]{aes}}
Components of the list are either quosures or constants.
}
\description{
\code{get_mappings} returns the mappings used by a ggplot object or a single
layer extracted from the object with \code{\link{get_geom_layer}} or \code{\link{get_stat_layer}}.
}
\details{
When passed a ggplot object (i.e. a plot), \code{get_mappings} will return
only the mappings that have been set globally with
\code{\link[ggplot2]{ggplot}}. When passed a single layer from a plot, the
behavior of \code{get_mappings} will depend on the value of
\code{local_only}. If \code{local_only = TRUE}, \code{get_mappings} will
return only the mappings defined locally in a layer. When \code{local_only =
FALSE}, \code{get_mappings} will return the combination of global and local
methods that will be used to plot a layer.
}
\examples{
require(ggplot2)
p <- ggplot(data = mpg, mapping = aes(x = displ, y = hwy)) +
  geom_point(mapping = aes(color = class))
get_mappings(p)
get_mappings(get_geom_layer(p, i = 1), local_only = FALSE)
}
\seealso{
Other functions for checking mappings: 
\code{\link{identical_aes}()},
\code{\link{ith_mappings_use}()},
\code{\link{ith_mappings}()},
\code{\link{uses_mappings}()}
}
\concept{functions for checking mappings}
