% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{get_stat_layer}
\alias{get_stat_layer}
\title{Isolate a stat layer from a plot}
\usage{
get_stat_layer(p, stat = NULL, i = NULL)
}
\arguments{
\item{p}{A ggplot object}

\item{stat}{A character string found in the suffix of a ggplot2 stat function,
e.g. \code{"bin"}.}

\item{i}{A numerical index, e.g. \code{1}.}
}
\value{
An object with class \code{layer_to_check} to be manipulated further
with ggcheck functions.
}
\description{
\code{get_stat_layer} returns a stat layer from a plot along with the global data sets
and aesthetic mappings that the layer may inherit from.
}
\details{
Users can specify a layer in one of 3 ways:

\enumerate{
\item By order of appearance with \code{i}. The first layer to appear in the
plot (the one drawn first, on the bottom) corresponds to \code{i = 1}.
\item By type of stat with \code{stat}. \code{get_stat_layer} will return the
first layer that uses the stat
\item By a combination of \code{stat} and
\code{i}. \code{get_stat_layer} will return the ith layer that uses the stat
}
}
\examples{
require(ggplot2)
p <- ggplot(data = diamonds, aes(price)) +
  stat_bin(bins = 20, binwidth = 500)

get_stat_layer(p, i = 1)
get_stat_layer(p, stat = "bin")
}
