% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_stats}
\alias{get_stats}
\title{List the stats used by a plot}
\usage{
get_stats(p)
}
\arguments{
\item{p}{A ggplot object}
}
\value{
A vector of character strings. Each element corresponds to the suffix
of a ggplot2 \code{stat_} function, e.g. \code{c("identity", "smooth")}.
}
\description{
\code{get_stats} returns a vector of stats names, written as character
strings, that describes which stats in which order are used by a plot.
}
\examples{
require(ggplot2)
p <- ggplot(data = mpg, mapping = aes(x = displ, y = hwy)) +
  geom_point(mapping = aes(color = class)) +
  geom_smooth()
get_stats(p)
}
\seealso{
Other functions for checking stats: 
\code{\link{ith_stat_is}()},
\code{\link{ith_stat}()},
\code{\link{uses_stats}()}
}
\concept{functions for checking stats}
