% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcheck-package.R
\docType{package}
\name{ggcheck-package}
\alias{ggcheck}
\alias{ggcheck-package}
\title{ggcheck: Inspect 'ggplot2' Plots for Automated Grading in Learning Exercises}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

'ggcheck' provides functions that inspect 'ggplot2' objects to make it easier for teachers to check that student plots meet expectations. Designed primarily for automated grading via 'gradethis' in interactive 'learnr' tutorials.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/ggcheck}
  \item Report bugs at \url{https://github.com/rstudio/ggcheck/issues}
}

}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{garrick@rstudio.com} (\href{https://orcid.org/0000-0002-7111-0077}{ORCID})

Authors:
\itemize{
  \item Garrett Grolemund \email{garrett@rstudio.com} (\href{https://orcid.org/0000-0002-7765-6011}{ORCID}) [conceptor]
  \item Nischal Shrestha \email{nsrocker92@gmail.com} (\href{https://orcid.org/0000-0003-3321-1712}{ORCID})
}

Other contributors:
\itemize{
  \item RStudio, PBC [copyright holder, funder]
}

}
\keyword{internal}
