% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grade_plot}
\alias{grade_plot}
\title{Grade a ggplot object with \code{gradethis::grade_result} conditions.}
\usage{
grade_plot(p_chunk, p, ...)
}
\arguments{
\item{p_chunk}{Source text for the ggplot code}

\item{p}{A ggplot object}

\item{...}{1 or more \code{gradethis_condition}}
}
\description{
Grade a ggplot object with \code{gradethis::grade_result} conditions.
}
