% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradethis_equal.R
\name{gradethis_equal.ggplot}
\alias{gradethis_equal.ggplot}
\title{Compare two \code{ggplot}s to check whether they are equal}
\usage{
\method{gradethis_equal}{ggplot}(x, y, ...)
}
\arguments{
\item{x, y}{Two \code{ggplot} objects to compare}

\item{...}{Unused}
}
\value{
A \link{logical} value of length one, or an internal gradethis error.
}
\description{
Compare two \code{ggplot}s to check whether they are equal
}
\examples{
library(ggplot2)

cty_plot <- ggplot(mpg, aes(x = displ, y = cty)) + geom_point()
hwy_plot <- ggplot(mpg, aes(x = displ, y = cty)) + geom_point()

gradethis_equal(cty_plot, hwy_plot)
gradethis_equal(cty_plot, cty_plot)
}
\seealso{
\code{\link[gradethis:gradethis_equal]{gradethis::gradethis_equal()}} for the generic function.
}
