% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{identical_aes}
\alias{identical_aes}
\title{Are aesthetic mapping specifications "identical"?}
\usage{
identical_aes(a1, a2)
}
\arguments{
\item{a1}{The output of \code{\link[ggplot2]{aes}}, perhaps extracted from a ggplot object.}

\item{a2}{The output of \code{\link[ggplot2]{aes}}, perhaps extracted from a ggplot object.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
The ggplot2 package uses quosures to record aesthetic mappings. These record
both the mapping described as well as the environment in which the mapping
was described. As a result, it is difficult to compare mappings created by
students in one environment to mappings created on the fly by graders in
another environment. \code{identical_aes} facilitates comparison by ignoring
the environments associated with an aesthetic mapping specification. If the
two specifications contain identical expressions, e.g. \code{x = displ},
etc., \code{identical_aes} returns \code{TRUE}.
}
\seealso{
Other functions for checking mappings: 
\code{\link{get_mappings}()},
\code{\link{ith_mappings_use}()},
\code{\link{ith_mappings}()},
\code{\link{uses_mappings}()}
}
\concept{functions for checking mappings}
