% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_ggplot.R
\name{is_ggplot}
\alias{is_ggplot}
\title{Check if an object is a ggplot}
\usage{
is_ggplot(p)
}
\arguments{
\item{p}{An object}
}
\value{
\code{\link{TRUE}} if \code{p} is a \link[ggplot2:ggplot]{ggplot} object,
otherwise \code{\link{FALSE}}
}
\description{
Check if an object is a ggplot
}
\examples{
require(ggplot2)

p_valid <- ggplot(data = mpg, mapping = aes(x = displ, y = hwy)) +
  geom_point()
is_ggplot(p_valid)

p_invalid <- geom_point()
is_ggplot(p_invalid)

}
