% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ith_data}
\alias{ith_data}
\title{Which data set does the ith layer use?}
\usage{
ith_data(p, i, local_only = FALSE)
}
\arguments{
\item{p}{A ggplot object or a layer extracted from a ggplot object with
\code{\link{get_layer}}.}

\item{i}{A numerical index that corresponds to the first layer of a plot (1),
the second layer (2), and so on.}

\item{local_only}{\code{TRUE} or \code{FALSE}. See the details.}
}
\value{
A data frame. If no data set is found, \code{ith_data} returns \code{NULL}.
}
\description{
\code{ith_data} returns the data set used by the ith layer.
}
\details{
If \code{local_only = TRUE}, \code{ith_data} returns the data set,
if any, that was defined locally in the function that created the ith layer.
If \code{local_only = FALSE}, \code{ith_data} returns the data used by
the ith layer, whether or not that data was defined globally in
\code{\link[ggplot2]{ggplot}} or locally.

Functions that use the \code{ith_} prefix are designed to eliminate the need
to call \code{get_layer} to check a specific layer in a plot, e.g. \code{p
\%>\% get_layer(geom = "point") \%>\% get_data()}.
}
\examples{
p <- ggplot(data = mpg, mapping = aes(x = displ, y = hwy)) +
  geom_point(data = d2, color = "red") +
  geom_point()
ith_data(p, i = 1)
}
\seealso{
Other functions for checking data: 
\code{\link{get_data}()},
\code{\link{ith_data_is}()},
\code{\link{uses_data}()}
}
\concept{functions for checking data}
