% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ith_data_is}
\alias{ith_data_is}
\title{Does the ith layer use the correct data set?}
\usage{
ith_data_is(p, data, i, local_only = FALSE)
}
\arguments{
\item{p}{A ggplot object or a layer extracted from a ggplot object with
\code{\link{get_layer}}.}

\item{data}{A data frame}

\item{i}{A numerical index that corresponds to the first layer of a plot (1),
the second layer (2), and so on.}

\item{local_only}{\code{TRUE} or \code{FALSE}. See the details.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
\code{ith_data_is} checks whether the student uses the supplied data set for
the ith layer of their plot.
}
\details{
Functions that use the \code{ith_} prefix are designed to eliminate the need
to call \code{get_layer} to check a specific layer in a plot, e.g. \code{p
\%>\% get_layer(geom = "point") \%>\% uses_data(mpg)}.

If \code{local_only = TRUE}, \code{ith_data_is} will check only the data set,
if any, that was defined locally in the function that created the ith layer.
If \code{local_only = FALSE}, \code{ith_data_is} will check the data used by
the ith layer, whether or not that data was defined globally in
\code{\link[ggplot2]{ggplot}} or locally.
}
\examples{
require(ggplot2)
d2 <- head(mpg)
p <- ggplot(data = mpg, mapping = aes(x = displ, y = hwy)) +
  geom_point(data = d2, color = "red") +
  geom_point()
ith_data_is(p, data = head(mpg), i = 1)
}
\seealso{
Other functions for checking data: 
\code{\link{get_data}()},
\code{\link{ith_data}()},
\code{\link{uses_data}()}
}
\concept{functions for checking data}
