% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{ith_geom}
\alias{ith_geom}
\title{Which geom is used in the ith layer?}
\usage{
ith_geom(p, i)
}
\arguments{
\item{p}{A ggplot object}

\item{i}{A numerical index that corresponds to the first layer of a plot (1),
the second layer (2), and so on.}
}
\value{
A character string that corresponds to the suffix of a ggplot2
\code{geom_} function, e.g. \code{"point"}.
}
\description{
\code{ith_geom} returns the type of geom used by the ith layer.
}
\examples{
p <- ggplot(data = mpg, mapping = aes(x = displ, y = hwy)) +
  geom_point(mapping = aes(color = class)) +
  geom_smooth()
ith_geom_is(p, geom = "smooth", i = 2)
}
\seealso{
Other functions for checking geoms: 
\code{\link{geoms_match}()},
\code{\link{get_geoms}()},
\code{\link{ith_geom_is}()},
\code{\link{uses_geoms}()}
}
\concept{functions for checking geoms}
