% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{ith_geom_is}
\alias{ith_geom_is}
\title{Is the ith geom what it should be?}
\usage{
ith_geom_is(p, geom, i = 1)
}
\arguments{
\item{p}{A ggplot object}

\item{geom}{A character string that corresponds to
the suffix of a ggplot2 \code{geom_} function, e.g. \code{"point"}.}

\item{i}{A numerical index that corresponds to the first layer of a plot (1),
the second layer (2), and so on. \code{ith_geom_is} will check the
geom used by the ith layer.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
\code{ith_geom_is} checks whether the ith layer uses the prescribed type of geom.
}
\examples{
require(ggplot2)
p <- ggplot(data = mpg, mapping = aes(x = displ, y = hwy)) +
  geom_point(mapping = aes(color = class)) +
  geom_smooth()
ith_geom_is(p, geom = "smooth", i = 2)
}
\seealso{
Other functions for checking geoms: 
\code{\link{geoms_match}()},
\code{\link{get_geoms}()},
\code{\link{ith_geom}()},
\code{\link{uses_geoms}()}
}
\concept{functions for checking geoms}
