% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{ith_geom_stat}
\alias{ith_geom_stat}
\title{Which geom/stat combination is used in the ith layer?}
\usage{
ith_geom_stat(p, i)
}
\arguments{
\item{p}{A ggplot object}

\item{i}{A numerical index that corresponds to the first layer of a plot (1),
the second layer (2), and so on.}
}
\value{
A list of lists with a GEOM and STAT strings, each corresponding to the suffix of a ggplot2
\code{geom_} function (e.g. \code{"point"}), and  \code{stat_} function (e.g. \code{"identity"}).
e.g. list(list(GEOM = "point", STAT = "identity"))
}
\description{
\code{ith_geom_stat} returns the type of geom used by the ith layer
according to a geom/stat combination.
}
\examples{
require(ggplot2)
p <- ggplot(data = mpg, mapping = aes(x = displ, y = hwy)) +
  geom_point(mapping = aes(color = class)) +
  geom_smooth()
ith_geom_stat(p, i = 2)
}
\seealso{
Other functions for checking geoms: 
\code{\link{get_geoms_stats}()},
\code{\link{get_geoms}()},
\code{\link{ith_geom_is}()},
\code{\link{ith_geom}()},
\code{\link{uses_geoms}()}
}
\concept{functions for checking geoms}
