% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{ith_mappings}
\alias{ith_mappings}
\title{Return the aesthetic mappings used by the ith layer}
\usage{
ith_mappings(p, i, local_only = FALSE)
}
\arguments{
\item{p}{A ggplot object or a layer extracted from a ggplot object with
\code{\link{get_layer}}.}

\item{i}{A numerical index that corresponds to the first layer of a plot (1),
the second layer (2), and so on. \code{ith_mappings_use} will check the
aesthetics used by the ith layer.}

\item{local_only}{If \code{TRUE}, \code{ith_mappings_use} will check only the
mappings defined locally in a layer for the presence of \code{mappings}. If
\code{FALSE}, \code{ith_mappings_use} will check for \code{mappings} in the
combination of global and local methods that will be used to plot a layer.}
}
\value{
A list with class uneval, as returned by \code{\link[ggplot2]{aes}}
Components of the list are either quosures or constants.
}
\description{
\code{ith_mappings} returns the mappings used by a ggplot object or a single
layer extracted from the object with \code{\link{get_layer}}.
}
\details{
Functions that use the \code{ith_} prefix are
designed to eliminate the need to call \code{get_layer} to check a specific
layer in a plot, e.g. \code{p \%>\% get_layer(geom = "point") \%>\% get_mappings()}.
}
\examples{
require(ggplot2)
p <- ggplot(data = mpg, mapping = aes(x = displ, y = hwy)) +
  geom_point(mapping = aes(color = class)) +
  geom_smooth()
ith_mappings(p, i = 1, local_only = FALSE)
ith_mappings(p, i = 1, local_only = TRUE)
ith_mappings(p, i = 2, local_only = FALSE)
}
\seealso{
Other functions for checking mappings: 
\code{\link{get_mappings}()},
\code{\link{identical_aes}()},
\code{\link{ith_mappings_match}()},
\code{\link{ith_mappings_use}()},
\code{\link{mappings_match}()},
\code{\link{uses_mappings}()}
}
\concept{functions for checking mappings}
