% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{ith_mappings_use}
\alias{ith_mappings_use}
\title{Does the ith layer use one or more aesthetic mappings?}
\usage{
ith_mappings_use(p, mappings, i, local_only = FALSE, exact = FALSE)
}
\arguments{
\item{p}{A ggplot object or a layer extracted from a ggplot object with
\code{\link{get_geom_layer}} or \code{\link{get_stat_layer}}.}

\item{mappings}{One or more aesthetic mappings created with
\code{\link[ggplot2]{aes}}.}

\item{i}{A numerical index that corresponds to the first layer of a plot (1),
the second layer (2), and so on. \code{ith_mappings_use} will check the
aesthetics used by the ith layer.}

\item{local_only}{If \code{TRUE}, \code{ith_mappings_use} will check only the
mappings defined locally in a layer for the presence of \code{mappings}. If
\code{FALSE}, \code{ith_mappings_use} will check for \code{mappings} in the
combination of global and local methods that will be used to plot a layer.}

\item{exact}{If \code{TRUE}, mappings need to be mapped exactly}
}
\value{
A logical value
}
\description{
\code{ith_mappings_use} checks whether the student uses the supplied mappings
in the ith layer of their plot.
}
\details{
\code{ith_mappings_use} ignores whether or not the student supplied
additional mappings as well. Functions that use the \code{ith_} prefix are
designed to eliminate the need to call \code{get_layer} to check a specific
layer in a plot, e.g. \code{p %>% get_geom_layer(geom = "point") %>%
uses_mappings(aes(color = class))}.
}
\examples{
require(ggplot2)
p <- ggplot(data = mpg, mapping = aes(x = displ, y = hwy)) +
  geom_point(mapping = aes(color = class)) +
  geom_smooth()
ith_mappings_use(p, i = 1, aes(x = displ), local_only = FALSE)
ith_mappings_use(p, i = 1, aes(x = displ), local_only = TRUE)
ith_mappings_use(p, i = 2, aes(x = displ, y = hwy), local_only = FALSE)
}
\seealso{
Other functions for checking mappings: 
\code{\link{get_mappings}()},
\code{\link{identical_aes}()},
\code{\link{ith_mappings}()},
\code{\link{uses_mappings}()}
}
\concept{functions for checking mappings}
