% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ith_stat}
\alias{ith_stat}
\title{Which stat is used in the ith layer?}
\usage{
ith_stat(p, i)
}
\arguments{
\item{p}{A ggplot object}

\item{i}{A numerical index that corresponds to the first layer of a plot (1),
the second layer (2), and so on.}
}
\value{
A character string that corresponds to the suffix of a ggplot2
\code{stat_} function, e.g. \code{"qq"}.
}
\description{
\code{ith_stat} returns the type of stat used by the ith layer.
}
\examples{
require(ggplot2)
p <- ggplot(data = diamonds, aes(sample = price)) +
  geom_qq()
ith_stat(p, i = 1)
}
\seealso{
Other functions for checking stats: 
\code{\link{get_stats}()},
\code{\link{uses_stats}()}
}
\concept{functions for checking stats}
