% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ith_stat_is}
\alias{ith_stat_is}
\title{Is the ith stat what it should be?}
\usage{
ith_stat_is(p, stat, i = 1)
}
\arguments{
\item{p}{A ggplot object}

\item{stat}{A character string that corresponds to
the suffix of a ggplot2 \code{stat_} function, e.g. \code{"identity"}.}

\item{i}{A numerical index that corresponds to the first layer of a plot (1),
the second layer (2), and so on. \code{ith_stat_is} will check the
stat used by the ith layer.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
\code{ith_stat_is} checks whether the ith layer uses the prescribed type of stat
}
\examples{
require(ggplot2)
p <- ggplot(data = diamonds, aes(sample = price)) +
  geom_qq()
ith_stat_is(p, i = 1, "qq")
}
\seealso{
Other functions for checking stats: 
\code{\link{get_stats}()},
\code{\link{ith_stat}()},
\code{\link{uses_stats}()}
}
\concept{functions for checking stats}
