% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{mappings_match}
\alias{mappings_match}
\title{Do the mappings of a plot or layer exactly match the set supplied?}
\usage{
mappings_match(p, mappings, local_only = FALSE)
}
\arguments{
\item{p}{A ggplot object or a layer extracted from a ggplot object with
\code{\link{get_layer}}.}

\item{mappings}{One or more aesthetic mappings created with
\code{\link[ggplot2]{aes}}.}

\item{local_only}{If \code{TRUE}, \code{mappings_match} will check only the
mappings defined locally in a layer for the presence of \code{mappings}. If
\code{FALSE}, \code{mappings_match} will check for \code{mappings} in the
combination of global and local methods that will be used to plot a layer.}
}
\value{
A logical value
}
\description{
\code{mappings_match} checks whether the student used the exact set of
supplied mappings in their plot.
}
\examples{
p <- ggplot(data = mpg, mapping = aes(x = displ, y = hwy)) +
  geom_point(mapping = aes(color = class))
mappings_match(p, aes(x = displ, y = hwy))
mappings_match(get_layer(p, i = 1), aes(x = displ,  y = hwy, color = class), local_only = FALSE)
mappings_match(get_layer(p, i = 1), aes(x = displ, y = hwy, color = class), local_only = TRUE)
}
\seealso{
Other check mappings: 
\code{\link{get_mappings}()},
\code{\link{ith_mappings_match}()},
\code{\link{ith_mappings_use}()},
\code{\link{ith_mappings}()},
\code{\link{uses_mappings}()}
}
\concept{check mappings}
