% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{n_layers}
\alias{n_layers}
\title{How many layers are in a plot?}
\usage{
n_layers(p)
}
\arguments{
\item{p}{A ggplot object}
}
\value{
Numeric. The number of layers.
}
\description{
How many layers are in a plot?
}
\examples{
p <- ggplot(data = mpg, mapping = aes(x = displ, y = hwy)) +
  geom_point(mapping = aes(color = class)) +
  geom_smooth()
n_layers(p)
}
