% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{uses_aesthetics}
\alias{uses_aesthetics}
\title{Does a plot use one or more aesthetics?}
\usage{
uses_aesthetics(p, aesthetics, local_only = FALSE, exact = FALSE)
}
\arguments{
\item{p}{A ggplot object or a layer extracted from a ggplot object with
\code{\link{get_layer}}.}

\item{aesthetics}{character vector of variables to check for, e.g. "x" or c("x")}

\item{local_only}{\code{TRUE} or \code{FALSE}. Should \code{uses_aesthetics} only
return mappings defined locally in the layer?}

\item{exact}{If \code{TRUE}, variables need to be mapped exactly}
}
\value{
A logical value.
}
\description{
\code{uses_aesthetics} checks whether the student used one or more aesthetics.
}
\details{
By default, \code{uses_aesthetics} requires that only one of the
aesthetics need to be used. Set \code{exact} to \code{TRUE} to check if all of
the variables have to be matched exactly.
}
