% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{uses_coordinate_system}
\alias{uses_coordinate_system}
\title{Does a plot use the correct coordinate system?}
\usage{
uses_coordinate_system(p, coordinates)
}
\arguments{
\item{p}{A ggplot2 object}

\item{coordinates}{A character string that corresponds to the suffix of a
ggplot2 \code{coord_} function, e.g. \code{"cartesian"}.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
\code{uses_coordinate_system} checks whether a plot uses the coordinate
system you describe. To describe a coordinate system, use the character
string that matches the suffix of the ggplot2 \code{coord_} function that
would make the coordinate system. The default coordinate system for ggplot2
plots is \code{"cartesian"}.
}
\examples{
require(ggplot2)
p <- ggplot(data = mpg, mapping = aes(x = displ, y = hwy)) +
  geom_point(mapping = aes(color = class)) +
  geom_smooth() +
  coord_polar()
uses_coordinate_system(p, coordinates = "polar")
}
\seealso{
Other functions for checking coordinate systems: 
\code{\link{get_coordinate_system}()}
}
\concept{functions for checking coordinate systems}
