% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{uses_data}
\alias{uses_data}
\title{Does a plot or layer use the correct data set?}
\usage{
uses_data(p, data, local_only = FALSE)
}
\arguments{
\item{p}{A ggplot object or a layer extracted from a ggplot object with
\code{\link{get_layer}}.}

\item{data}{A data frame}

\item{local_only}{\code{TRUE} or \code{FALSE}. See the details.}
}
\value{
A data frame.
}
\description{
\code{uses_data} checks whether the data set used by a plot or layer matches
the data set provided.
}
\details{
When passed a ggplot object (i.e. a plot), \code{uses_data} will check only
the data that has been set globally with \code{\link[ggplot2]{ggplot}}.

When passed a single layer from a plot, the behavior of \code{uses_data} will
depend on the \code{local_only} argument passed to \code{...}. If
\code{local_only = TRUE}, \code{uses_data} will check only the data set, if
any, that was defined locally in the function that created the layer. If
\code{local_only = FALSE}, \code{uses_data} will check the data used by the
layer, whether or not that data was defined globally in
\code{\link[ggplot2]{ggplot}} or locally.
}
\examples{
p <- ggplot(data = mpg, mapping = aes(x = displ, y = hwy)) +
  geom_point(data = d2, color = "red") +
  geom_point()
uses_data(p, mpg)
uses_data(get_layer(p, i = 1), data = head(mpg))
}
\seealso{
Other functions for checking data: 
\code{\link{get_data}()},
\code{\link{ith_data_is}()},
\code{\link{ith_data}()}
}
\concept{functions for checking data}
