% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{uses_extra_mappings}
\alias{uses_extra_mappings}
\title{Does the plot uses extra aesthetic mappings?}
\usage{
uses_extra_mappings(p, mappings, local_only = FALSE)
}
\arguments{
\item{p}{A ggplot object or a layer extracted from a ggplot object with
\code{\link{get_geom_layer}} or \code{\link{get_stat_layer}}.}

\item{mappings}{One or more aesthetic mappings created with
\code{\link[ggplot2]{aes}}.}

\item{local_only}{If \code{TRUE}, \code{uses_extra_mappings} will check only the
mappings defined locally in a layer for the presence of \code{mappings}. If
\code{FALSE}, \code{uses_extra_mappings} will check for \code{mappings} in the
combination of global and local methods that will be used to plot a layer.}
}
\value{
A logical value.
}
\description{
\code{uses_extra_mappings} checks if a student's plot contains more than the
required aesthetic mappings. Note that we still return \code{TRUE} if
the student's plot differs from the required aesthetic mappings because they
are technically extra mappings from required set. We recommend you use
\code{uses_mapping} checks for checking required mappings before \code{uses_extra_mappings}.
}
\examples{
require(ggplot2)
p <- ggplot(data = diamonds, aes(x = cut, sample = price)) +
  geom_qq()
uses_extra_mappings(p, aes(sample = price))
}
