% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{uses_geom_param}
\alias{uses_geom_param}
\title{Does a layer use a specific geom parameter?}
\usage{
uses_geom_param(p, geom, param, i = NULL)
}
\arguments{
\item{p}{A ggplot object}

\item{geom}{A character string found in the suffix of a ggplot2 geom function,
e.g. \code{"point"}.}

\item{param}{A named list, e.g. list(outlier.alpha)}

\item{i}{A numerical index, e.g. \code{1}.}
}
\value{
A boolean
}
\description{
\code{uses_geom_param} checks that a plot's geom layer uses a specific geom parameter.
}
\details{
To specify a specific geom layer, either specify the position using the \code{i} index or
by a combination of \code{geom} function suffix name and \code{i} to check the ith layer that
uses the geom.
}
\examples{
require(ggplot2)
p <- ggplot(data = diamonds, aes(x = cut, y = price)) +
  geom_boxplot(outlier.alpha = 0.01)
uses_geom_param(p, geom = "boxplot", param = list(outlier.alpha = 0.01))
}
