% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_params.R
\name{uses_geom_params}
\alias{uses_geom_params}
\alias{uses_geom_param}
\title{Does a layer use one of more specific parameters?}
\usage{
uses_geom_params(p, geom, ..., params = NULL, i = NULL)

uses_geom_param(p, geom, ..., params = NULL, i = NULL)
}
\arguments{
\item{p}{A ggplot object}

\item{geom}{A character string found in the suffix of a ggplot2 geom function,
e.g. \code{"point"}.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}>
Named values or \link{character} strings.
Unnamed arguments will check whether any value was set for that parameter.
Named arguments will check whether the parameter with the same name has a
matching value.
Each argument should have a name matching a \link[ggplot2:ggplot]{ggplot}
layer parameter.
Values may be passed as arguments or as list elements.}

\item{params}{A named list of geom or stat parameter values, e.g.
\code{list(outlier.alpha = 0.01)}.
This list is combined with any inputs to \code{...}}

\item{i}{A numerical index, e.g. \code{1}.}
}
\value{
A named logical vector of the same length as the number of inputs
to \code{...}.
}
\description{
\code{uses_geom_params} checks that a plot's geom layer uses a specific parameter.
}
\details{
To specify a specific geom layer, either specify using position using the \code{i} index or
by using a combination of \code{geom} function suffix name and \code{i} to check the ith layer that
uses the geom.

The \code{params} argument accepts a list that contains geom, stat, or aes
parameters. This offers flexibility in certain situations where setting a
parameter on a \code{geom_} function is actually setting a stat parameter or
aes parameter. For example, in \code{geom_histogram(binwidth = 500)}, the
\code{binwidth} is a stat parameter, while in
\code{geom_histogram(fill = "blue")}, the \code{fill} is an aes parameter.
\code{uses_geom_params} will take this into account and check geom, stat, and
aes parameters.

Note that \code{uses_geom_params()} can detect aes \emph{parameters}, but not aes
\emph{mappings}. Parameters are set directly within a layer (e.g.
\code{geom_point(color = "blue")}), while mappings are set using
\code{\link[ggplot2:aes]{aes()}} (e.g. \code{geom_point(aes(color = class))}).
}
\examples{
require(ggplot2)

p <- ggplot(data = diamonds, aes(x = cut, y = price)) +
  geom_boxplot(varwidth = TRUE, outlier.alpha = 0.01, fill = "blue")

uses_geom_params(
  p, "boxplot", list(varwidth = TRUE, outlier.alpha = 0.01, fill = "blue")
)

uses_geom_params(
  p, "boxplot", varwidth = TRUE, outlier.alpha = 0.01, fill = "blue"
)

# Unnamed arguments check that a parameter is set to any value
uses_geom_params(p, "boxplot", "fill")
}
\seealso{
Other functions for checking geom parameters: 
\code{\link{default_params}()}
}
\concept{functions for checking geom parameters}
