% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{uses_stat_param}
\alias{uses_stat_param}
\title{Does a layer use a specific stat parameter?}
\usage{
uses_stat_param(p, stat, params, i = NULL)
}
\arguments{
\item{p}{A ggplot object}

\item{stat}{A character string found in the suffix of a ggplot2 stat function,
e.g. \code{"bin"}.}

\item{params}{A named list of stat or geom parameter values, e.g. \code{list(bins = 200)}}

\item{i}{A numerical index, e.g. \code{1}.}
}
\value{
A boolean
}
\description{
\code{uses_stat_param} is a mirror function of \code{uses_geom_param} but instead of checking a plot's
geom layer, it checks that a plot's stat layer uses a specific stat parameter.
}
\details{
To specify a specific stat layer, either specify using position using the \code{i} index or
by using a combination of \code{stat} function suffix name and \code{i} to check the ith layer that
uses the stat.
}
\examples{
require(ggplot2)
p <- ggplot(diamonds, aes(carat)) +
  stat_bin(bins = 200)
uses_stat_param(p, stat = "bin", params = list(bins = 200))
}
