% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{uses_stats}
\alias{uses_stats}
\title{Does a plot use one or more stats?}
\usage{
uses_stats(p, stats, geoms = NULL, exact = TRUE)
}
\arguments{
\item{p}{A ggplot object}

\item{stats}{A vector of character strings. Each element should correspond to
the suffix of a ggplot2 \code{stat_} function, e.g. \code{c("identity", "smooth")}.}

\item{geoms}{A character vector to optionally check for the geoms corresponding to stats
e.g. c("point", "smooth") if checking c("identity", "smooth")}

\item{exact}{if \code{TRUE}, use exact matching}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
\code{uses_stats} tests whether a plot uses one or more stats in its layers.
}
\details{
By default, the plot must have the exact stats or geom/stat combinations and in the same order.
However, if \code{exact} is set to \code{FALSE}, the plot stats or geom/stat combinations do not have to be exact.
}
\examples{
require(ggplot2)
p <- ggplot(data = mpg, mapping = aes(x = displ, y = hwy)) +
  geom_point(mapping = aes(color = class)) +
  geom_smooth()
uses_stats(p, stats = "smooth")
uses_stats(p, stats = c("identity", "smooth"), exact = TRUE)
uses_stats(p, c("smooth", "identity"), geoms = c("smooth", "point"))
}
\seealso{
Other functions for checking stats: 
\code{\link{get_stats}()},
\code{\link{ith_stat_is}()},
\code{\link{ith_stat}()}
}
\concept{functions for checking stats}
