% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{uses_variables}
\alias{uses_variables}
\title{Does a plot use one or more variables?}
\usage{
uses_variables(p, vars, i = 1, exact = FALSE)
}
\arguments{
\item{p}{A ggplot object or a layer extracted from a ggplot object with
\code{\link{get_layer}}.}

\item{vars}{character vector of variables to check for, e.g. c("x")}

\item{i}{the ith layer to check}

\item{exact}{If \code{TRUE}, variables need to be mapped exactly}
}
\value{
A logical value.
}
\description{
TODO-Nischal need to fix error when running example, suspicion: get_layer breaks because it
is relying on get_geom and is not aware of get_stats cases. Might be worth rethinking the
entire geom_ vs stat_ dynamics.
}
\details{
\code{uses_variables} checks whether the student used one or more variables in
their plot aesthetics. By default, \code{uses_variables} requires that only one of the
variables need to be used. Set \code{exact} to \code{TRUE} to check if all of the variables
have to be used.
}
\examples{
require(ggplot2)
p <- ggplot(data = diamonds, aes(x = cut, sample = price)) +
  geom_qq()
uses_variables(p, c("x", "y"))
}
