% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_placeholders.R
\name{default_label}
\alias{default_label}
\alias{default_param}
\title{Placeholders for default values}
\usage{
default_label()

default_param()
}
\value{
A placeholder value to be used within \code{\link[=uses_labels]{uses_labels()}}
or \code{\link[=uses_geom_params]{uses_geom_params()}}.
}
\description{
These functions generate placeholder values.
\itemize{
\item \code{default_label()} can be used as a named argument in \code{\link[=uses_labels]{uses_labels()}}
to check that a label matches the result of \code{\link[=get_default_labels]{get_default_labels()}}
with that name.
\item \code{default_param()} can be used as a named argument in \code{\link[=uses_geom_params]{uses_geom_params()}}
to check that a parameter matched the result of \code{\link[=get_default_params]{get_default_params()}}
with that name.
}
}
\examples{
require(ggplot2)

p <- ggplot(data = mpg, mapping = aes(x = displ, y = hwy, color = trans)) +
  geom_smooth(se = FALSE) +
  labs(title = "My plot", x = "Weight", y = "MPG")

uses_labels(p, x = default_label(), color = default_label())

uses_geom_params(p, "smooth", size = default_param(), se = default_param())
}
