% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{uses_aesthetics}
\alias{uses_aesthetics}
\title{Does a plot use one or more aesthetics?}
\usage{
uses_aesthetics(p, aesthetics, local_only = FALSE, exact = FALSE)
}
\arguments{
\item{p}{A ggplot object or a layer extracted from a ggplot object with
\code{\link{get_geom_layer}} or \code{\link{get_stat_layer}}..}

\item{aesthetics}{character vector of variables to check for, e.g. "x" or c("x")}

\item{local_only}{\code{TRUE} or \code{FALSE}. Should \code{uses_aesthetics} only
return mappings defined locally in the layer?}

\item{exact}{If \code{TRUE}, variables need to be mapped exactly}
}
\value{
A logical value.
}
\description{
\code{uses_aesthetics} checks whether the student used one or more aesthetics.
}
\details{
By default, \code{uses_aesthetics} requires that only one of the
aesthetics need to be used. Set \code{exact} to \code{TRUE} to check if all of
the variables have to be matched exactly.
}
\examples{
require(ggplot2)
p <- ggplot(data = mpg, mapping = aes(x = displ, y = hwy)) +
  geom_point(mapping = aes(color = class))
uses_aesthetics(p, "x")
uses_aesthetics(p, c("x", "y"))
uses_aesthetics(get_geom_layer(p, "point"), c("x", "y", "color"), local_only = TRUE)
uses_aesthetics(get_geom_layer(p, "point"), c("x", "y"), local_only = FALSE)
}
