% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{uses_geom_param}
\alias{uses_geom_param}
\title{Does a layer use a specific parameter?}
\usage{
uses_geom_param(p, geom, ..., params = NULL, i = NULL)
}
\arguments{
\item{p}{A ggplot object}

\item{geom}{A character string found in the suffix of a ggplot2 geom function,
e.g. \code{"point"}.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Named values.
Each parameter should have a name matching a \link[ggplot2:ggplot]{ggplot}
layer parameter and a value matching the expected label.
Named values may be passed as arguments or as list elements.}

\item{params}{A named list of geom or stat parameter values, e.g.
\code{list(outlier.alpha = 0.01)}.
This list is combined with any inputs to \code{...}}

\item{i}{A numerical index, e.g. \code{1}.}
}
\value{
A boolean
}
\description{
\code{uses_geom_param} checks that a plot's geom layer uses a specific parameter.
}
\details{
To specify a specific geom layer, either specify using position using the \code{i} index or
by using a combination of \code{geom} function suffix name and \code{i} to check the ith layer that
uses the geom.

The \code{params} argument accepts a list that contains geom, stat, or aes
parameters. This offers flexibility in certain situations where setting a
parameter on a \code{geom_} function is actually setting a stat parameter or
aes parameter. For example, in \code{geom_histogram(binwidth = 500)}, the
\code{binwidth} is a stat parameter, while in
\code{geom_histogram(fill = "blue")}, the \code{fill} is an aes parameter.
\code{uses_geom_param} will take this into account and check geom, stat, and
aes parameters.
}
\examples{
require(ggplot2)
p <- ggplot(data = diamonds, aes(x = cut, y = price)) +
  geom_boxplot(varwidth = TRUE, outlier.alpha = 0.01)
uses_geom_param(p, geom = "boxplot", params = list(varwidth = TRUE, outlier.alpha = 0.01))
}
