% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{uses_labels}
\alias{uses_labels}
\title{Does a plot use one or more labels?}
\usage{
uses_labels(p, ...)
}
\arguments{
\item{p}{A ggplot object}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Named \link{character} strings.
Each argument should have a name matching a \link[ggplot2:ggplot]{ggplot}
\link[ggplot2:aes]{aesthetic} or \link[ggplot2:labs]{label},
and a value matching the expected label.
Named character strings may be passed as arguments or as list elements.}
}
\value{
A logical vector of the same length as the number of inputs to \code{...}.
}
\description{
\code{uses_labels()} tests whether a plot uses one or more \link[ggplot2:labs]{labels}.
}
\details{
Note that \code{uses_labels()} will match \code{\link{NULL}} if a label is explicitly set to
\code{\link{NULL}} \emph{\strong{or}} if a requested aesthetic is not present in the plot.
}
\examples{
require(ggplot2)

p <- ggplot(data = mpg, mapping = aes(x = displ, y = hwy)) +
  geom_point(mapping = aes(color = class)) +
  geom_smooth() +
  labs(x = "Weight", y = "MPG", color = NULL)

uses_labels(p, x = "Weight")
uses_labels(p, x = "Weight", y = "MPG", color = NULL)

# The colo(u)r aesthetic can be matched with or without a u
uses_labels(p, color = NULL)
uses_labels(p, colour = NULL)

# Inputs can be passed from a list, with or without the !!! operator
label_list <- list(x = "Weight", y = "MPG", color = NULL)
uses_labels(p, label_list)
uses_labels(p, !!!label_list)
}
\seealso{
Other functions for checking labels: 
\code{\link{get_labels}()}
}
\concept{functions for checking labels}
