% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{uses_labels}
\alias{uses_labels}
\title{Does a plot use one or more labels?}
\usage{
uses_labels(p, ...)
}
\arguments{
\item{p}{A ggplot object}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}>
\link[=character]{Character} strings.
Unnamed arguments will check whether a label exists for that aesthetic.
Named arguments will check whether the aesthetic with the same name
has a label with a matching value.
Each argument should have a matching \link[ggplot2:ggplot]{ggplot}
\link[ggplot2:aes]{aesthetic} or \link[ggplot2:labs]{label}.
Strings may be input as individual arguments or as list elements.}
}
\value{
A named logical vector of the same length as the number of inputs
to \code{...}.
}
\description{
\code{uses_labels()} tests whether a plot uses one or more \link[ggplot2:labs]{labels}.
}
\details{
Note that \code{uses_labels()} will match \code{\link{NULL}} if a label is explicitly set to
\code{\link{NULL}} \emph{\strong{or}} if a requested aesthetic is not present in the plot.
}
\examples{
require(ggplot2)

p <- ggplot(data = mpg, mapping = aes(x = displ, y = hwy)) +
  geom_point(mapping = aes(color = class, shape = drv)) +
  geom_smooth() +
  labs(title = "My plot", x = "Weight", y = "MPG", color = NULL)

# Unnamed arguments check if a label is set for the given aesthetic
uses_labels(p, "title", "subtitle", "x", "y")

# The check will return TRUE for labels set to NULL
uses_labels(p, "color")

# The check will return TRUE for aesthetics with default labels
uses_labels(p, "shape")

# Named arguments check if the label matches an expected value
uses_labels(p, x = "Weight")
uses_labels(p, x = "Weight", y = "MPG", color = NULL)

# You can check for default labels with default_label()
uses_labels(p, shape = default_label(), x = default_label())

# The colo(u)r aesthetic can be matched with or without a u
uses_labels(p, color = NULL)
uses_labels(p, colour = NULL)

# Inputs can be passed from a list, with or without the !!! operator
label_list <- list(x = "Weight", y = "MPG", color = NULL)
uses_labels(p, label_list)
uses_labels(p, !!!label_list)
}
\seealso{
Other functions for checking labels: 
\code{\link{get_default_labels}()},
\code{\link{get_labels}()}
}
\concept{functions for checking labels}
