% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{uses_mappings}
\alias{uses_mappings}
\title{Does a plot or layer use one or more mappings?}
\usage{
uses_mappings(p, mappings, local_only = FALSE)
}
\arguments{
\item{p}{A ggplot object or a layer extracted from a ggplot object with
\code{\link{get_layer}}.}

\item{mappings}{One or more aesthetic mappings created with
\code{\link[ggplot2]{aes}}.}

\item{local_only}{If \code{TRUE}, \code{uses_mappings} will check only the
mappings defined locally in a layer for the presence of \code{mappings}. If
\code{FALSE}, \code{uses_mappings} will check for \code{mappings} in the
combination of global and local methods that will be used to plot a layer.}
}
\value{
A logical value.
}
\description{
\code{uses_mappings} checks whether the student used one or more mappings in
their plot. Unlike, \code{\link{mappings_match}}, \code{use_mappings} ignores
whether or not the student also supplied additional mappings.
}
\examples{
p <- ggplot(data = mpg, mapping = aes(x = displ, y = hwy)) +
  geom_point(mapping = aes(color = class))
uses_mappings(p, aes(x = displ))
uses_mappings(get_layer(p, i = 1), aes(x = displ, color = class), local_only = FALSE)
uses_mappings(get_layer(p, i = 1), aes(x = displ, color = class), local_only = TRUE)
}
\seealso{
Other check mappings: 
\code{\link{get_mappings}()},
\code{\link{ith_mappings_match}()},
\code{\link{ith_mappings_use}()},
\code{\link{ith_mappings}()},
\code{\link{mappings_match}()}
}
\concept{check mappings}
