% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{uses_stats}
\alias{uses_stats}
\title{Does a plot use one or more stats?}
\usage{
uses_stats(p, stats, exact = FALSE)
}
\arguments{
\item{p}{A ggplot object}

\item{stats}{A vector of character strings. Each element should correspond to
the suffix of a ggplot2 \code{stat_} function, e.g. \code{c("identity", "smooth")}.}

\item{exact}{if \code{TRUE}, use exact matching}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
\code{uses_stats} tests whether a plot uses one or more stats in its layers.
The stats can appear in any order in the plot and can be accompanied by other
stats that are not checked for.
}
\examples{
require(ggplot2)
p <- ggplot(data = mpg, mapping = aes(x = displ, y = hwy)) +
  geom_point(mapping = aes(color = class)) +
  geom_smooth()
uses_stats(p, stats = "smooth")
}
\seealso{
Other functions for checking stats: 
\code{\link{get_stats}()},
\code{\link{ith_stat}()}
}
\concept{functions for checking stats}
