% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{uses_variables}
\alias{uses_variables}
\title{Does a plot use one or more variables?}
\usage{
uses_variables(p, vars, i = 1, local_only = FALSE)
}
\arguments{
\item{p}{A ggplot object or a layer extracted from a ggplot object with
\code{\link{get_layer}}.}

\item{vars}{character vector of variables to check for, e.g. c("x")}

\item{i}{the ith layer to check}

\item{local_only}{If \code{TRUE}, \code{uses_variables} will check only the
mappings defined locally in a layer for the presence of \code{mappings}. If
\code{FALSE}, \code{uses_variables} will check for \code{mappings} in the
combination of global and local methods that will be used to plot a layer.}
}
\value{

}
\description{
TODO-Nischal need to fix error when running example, suspicion: get_layer breaks because it
is relying on get_geom and is not aware of get_stats cases. Might be worth rethinking the
entire geom_ vs stat_ dynamics.
}
\details{
\code{uses_variables} checks whether the student used one or more variables in
their plot aesthetics. It is not an exact match, so it will return \code{TRUE}
if any of the variables are used.
}
\examples{
\dontrun{
require(ggplot2)
p <- ggplot(data = diamonds, aes(x = cut, sample = price))
  geom_qq()
uses_variables(p, c("x", "y"))
}
}
